python_server/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.31
mysql-connector-python==8.4.0
web3==6.19.0
google-generativeai==0.7.1
python-dotenv==1.0.1
pydantic==2.7.4
cors==1.0.1
